Add in the idt application, generators, from the broken lfs branch Ensure everything is passing the precommit hook checks
