+++ aces/idt/generators/__init__.py
@@ -1,0 +1,3 @@
+from .prosumer_camera import IDTGeneratorProsumerCamera
+
+ALL_GENERATORS = {IDTGeneratorProsumerCamera.generator_name: IDTGeneratorProsumerCamera}
