+++ tests/resources/samples_camera.json
@@ -1,0 +1,44 @@
+[
+  [0.11395355314016342, 0.11399932950735092, 0.11435028165578842],
+  [0.13511893153190613, 0.13519522547721863, 0.13589711487293243],
+  [0.15139985084533691, 0.15228484570980072, 0.15271209180355072],
+  [0.1743488907814026, 0.1744709610939026, 0.1755695790052414],
+  [0.19754229485988617, 0.19870196282863617, 0.19926652312278748],
+  [0.2040577530860901, 0.2042103409767151, 0.2039509415626526],
+  [0.22654908895492554, 0.22670167684555054, 0.22805969417095184],
+  [0.2544451951980591, 0.25578805804252625, 0.25645944476127625],
+  [0.25748175382614136, 0.25856509804725647, 0.25832095742225647],
+  [0.2620593011379242, 0.2622424066066742, 0.2619524896144867],
+  [0.28784626722335815, 0.28802937269210815, 0.28955525159835815],
+  [0.30214357376098633, 0.30269289016723633, 0.30197572708129883],
+  [0.31888240575790405, 0.32036247849464417, 0.32109490036964417],
+  [0.32222405076026917, 0.3233989477157593, 0.3231395483016968],
+  [0.3272288739681244, 0.3274119794368744, 0.3271068036556244],
+  [0.34091585874557495, 0.34119051694869995, 0.33760473132133484],
+  [0.3550606369972229, 0.3552437424659729, 0.356876403093338],
+  [0.3702734708786011, 0.3708533048629761, 0.3701056241989136],
+  [0.3879435658454895, 0.3894999325275421, 0.3902781307697296],
+  [0.3914530277252197, 0.3926889896392822, 0.3924143314361572],
+  [0.39668676257133484, 0.39688512682914734, 0.39656469225883484],
+  [0.4109841585159302, 0.4112740755081177, 0.4075356721878052],
+  [0.42567822337150574, 0.42587658762931824, 0.42755505442619324],
+  [0.44139474630355835, 0.44200509786605835, 0.44122692942619324],
+  [0.4595836102962494, 0.4611705541610718, 0.4619640111923218],
+  [0.4631688892841339, 0.46443599462509155, 0.46416133642196655],
+  [0.4685399830341339, 0.4687541723251343, 0.4684034585952759],
+  [0.48315778374671936, 0.48344892263412476, 0.47964832186698914],
+  [0.5141263604164124, 0.5147498846054077, 0.5139564275741577],
+  [0.5325718522071838, 0.5341892838478088, 0.5350005626678467],
+  [0.5362212657928467, 0.5375030040740967, 0.5372104048728943],
+  [0.5416533946990967, 0.5418670177459717, 0.5415313243865967],
+  [0.5564362406730652, 0.5567413568496704, 0.5528839230537415],
+  [0.5876858234405518, 0.5883151292800903, 0.5875216722488403],
+  [0.6099328398704529, 0.6112339496612549, 0.6109399199485779],
+  [0.6154148578643799, 0.6156089901924133, 0.6152732968330383],
+  [0.6302772760391235, 0.6305824518203735, 0.626706600189209],
+  [0.6616800427436829, 0.6622903943061829, 0.6614969372749329],
+  [0.6839984059333801, 0.6853007674217224, 0.6850054860115051],
+  [0.704404890537262, 0.704710066318512, 0.7008135318756104],
+  [0.7358686327934265, 0.7364881634712219, 0.7356947064399719],
+  [0.7786641716957092, 0.7789693474769592, 0.7750848531723022]
+]
