+++ tests/resources/samples_reference.json
@@ -1,0 +1,44 @@
+[
+  [0.004681354292352056, 0.004707711883419395, 0.004869287638457769],
+  [0.009362708584704112, 0.00941542376683879, 0.009738575276915538],
+  [0.011834001784287838, 0.011900426739566225, 0.012045619808943012],
+  [0.018725417169408225, 0.01883084753367758, 0.019477150553831077],
+  [0.023668003568575676, 0.02380085347913245, 0.024091239617886024],
+  [0.025315286747004714, 0.025303940415945837, 0.025357141282908682],
+  [0.03745083433881645, 0.03766169506735516, 0.038954301107662154],
+  [0.04733600713715135, 0.0476017069582649, 0.04818247923577205],
+  [0.044181965243637135, 0.04417079370539635, 0.0442352625071515],
+  [0.05063057349400943, 0.050607880831891675, 0.050714282565817365],
+  [0.0749016686776329, 0.07532339013471032, 0.07790860221532431],
+  [0.0716937369523808, 0.07156985933022324, 0.07145486681949595],
+  [0.0946720142743027, 0.0952034139165298, 0.0963649584715441],
+  [0.08836393048727427, 0.0883415874107927, 0.088470525014303],
+  [0.10126114698801886, 0.10121576166378335, 0.10142856513163473],
+  [0.10831335388455121, 0.10848966830035922, 0.107261735165307],
+  [0.1498033373552658, 0.15064678026942063, 0.15581720443064861],
+  [0.1433874739047616, 0.14313971866044647, 0.1429097336389919],
+  [0.1893440285486054, 0.1904068278330596, 0.1927299169430882],
+  [0.17672786097454854, 0.1766831748215854, 0.176941050028606],
+  [0.20252229397603771, 0.2024315233275667, 0.20285713026326946],
+  [0.21662670776910242, 0.21697933660071844, 0.214523470330614],
+  [0.2996066747105316, 0.30129356053884127, 0.31163440886129723],
+  [0.2867749478095232, 0.28627943732089295, 0.2858194672779838],
+  [0.3786880570972108, 0.3808136556661192, 0.3854598338861764],
+  [0.3534557219490971, 0.3533663496431708, 0.353882100057212],
+  [0.40504458795207543, 0.4048630466551334, 0.4057142605265389],
+  [0.43325341553820484, 0.4339586732014369, 0.429046940661228],
+  [0.5735498956190463, 0.5725588746417859, 0.5716389345559676],
+  [0.7573761141944216, 0.7616273113322384, 0.7709196677723528],
+  [0.7069114438981942, 0.7067326992863416, 0.707764200114424],
+  [0.8100891759041509, 0.8097260933102668, 0.8114285210530778],
+  [0.8665068310764097, 0.8679173464028738, 0.858093881322456],
+  [1.1470997912380927, 1.1451177492835718, 1.1432778691119352],
+  [1.4138228877963883, 1.4134653985726833, 1.415528400228848],
+  [1.6201783518083017, 1.6194521866205336, 1.6228570421061557],
+  [1.7330136621528194, 1.7358346928057475, 1.716187762644912],
+  [2.2941995824761854, 2.2902354985671436, 2.2865557382238704],
+  [2.8276457755927766, 2.8269307971453665, 2.831056800457696],
+  [3.4660273243056388, 3.471669385611495, 3.432375525289824],
+  [4.588399164952371, 4.580470997134287, 4.573111476447741],
+  [6.9320546486112775, 6.94333877122299, 6.864751050579648]
+]
