+++ tests/resources/synthetic_002/synthetic_002.json
@@ -7,11 +7,20 @@
     "camera_model": "colour-science",
     "iso": 100,
     "temperature": 6500,
-    "additional_camera_settings": null,
+    "additional_camera_settings": "",
     "lighting_setup_description": "D65",
-    "debayering_platform": null,
-    "debayering_settings": null,
-    "encoding_colourspace": null
+    "debayering_platform": "",
+    "debayering_settings": "",
+    "encoding_colourspace": "",
+    "rgb_display_colourspace": "sRGB",
+    "cat": "CAT02",
+    "optimisation_space": "Oklab",
+    "illuminant_interpolator": "Linear",
+    "decoding_method": "Median",
+    "ev_range": [-1.0, 0.0, 1.0],
+    "grey_card_reference": [0.18, 0.18, 0.18],
+    "lut_size": 1024,
+    "lut_smoothing": 32
   },
   "data": {
     "colour_checker": {
